refactor: move FileItem to app-file-manager
