+++ packages/app-file-manager/src/BatchFileUploader.ts
@@ -1,6 +1,6 @@
 import { generateId } from "@webiny/utils";
 import { makeAutoObservable, runInAction, action } from "mobx";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { FileManagerViewContext } from "~/modules/FileManagerRenderer/FileManagerViewProvider/index.js";
 
 export interface FileError {
