+++ packages/app-file-manager/src/components/BulkActions/ActionEdit/ActionEdit.types.ts
@@ -1,3 +1,3 @@
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 
 export type ActionFormData = Partial<Omit<FileItem, "id">>;
