+++ packages/app-file-manager/src/components/BulkActions/ActionEdit/GraphQLInputMapper.ts
@@ -1,6 +1,6 @@
 import set from "lodash/set.js";
 import get from "lodash/get.js";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { BatchDTO } from "~/components/BulkActions/ActionEdit/domain/index.js";
 import { OperatorType } from "~/components/BulkActions/ActionEdit/domain/index.js";
