+++ packages/app-file-manager/src/components/EditFileUsingUrl/GetFileByUrl.ts
@@ -1,6 +1,6 @@
 import type ApolloClient from "apollo-client";
 import type { DocumentNode } from "graphql";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { IGetFileByUrl } from "./EditFileUsingUrlRepository.js";
 import { getFileByUrlQuery } from "./getFileByUrl.graphql.js";
 import type { useFileModel } from "~/hooks/useFileModel.js";
