+++ packages/app-file-manager/src/components/EditFileUsingUrl/UpdateFile.ts
@@ -1,4 +1,4 @@
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { useFileManagerApi } from "~/modules/FileManagerApiProvider/FileManagerApiContext/index.js";
 import type { IUpdateFile } from "./EditFileUsingUrlRepository.js";
