+++ packages/app-file-manager/src/components/FileDetails/FileDetailsProvider.tsx
@@ -1,6 +1,6 @@
 import React, { createContext } from "react";
 import { useFileManagerView } from "~/modules/FileManagerRenderer/FileManagerViewProvider/index.js";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 
 export interface FileDetailsContext {
     close: () => void;
