+++ packages/app-file-manager/src/hooks/useCopyFile.tsx
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import { i18n } from "@webiny/app/i18n/index.js";
 import { useSnackbar } from "@webiny/app-admin";
 import { useFileManagerView } from "~/index.js";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 
 const t = i18n.ns("app-admin/file-manager/hooks/use-copy-file");
