+++ packages/app-file-manager/src/hooks/useDeleteFile.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react";
 import { i18n } from "@webiny/app/i18n/index.js";
 import { OverlayLoader, Text } from "@webiny/admin-ui";
 import { useConfirmationDialog, useSnackbar } from "@webiny/app-admin";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import { useFileManagerView } from "~/index.js";
 
 const t = i18n.ns("app-admin/file-manager/hooks/use-delete-file");
