+++ packages/app-file-manager/src/modules/FileManagerApiProvider/FileManagerApiContext/useFileManagerApi.ts
@@ -1,4 +1,4 @@
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import React from "react";
 import type { FileManagerApiContextData } from "./FileManagerApiContext.js";
 import { FileManagerApiContext } from "./FileManagerApiContext.js";
