+++ packages/app-file-manager/src/modules/FileManagerApiProvider/graphql.ts
@@ -1,5 +1,5 @@
 import gql from "graphql-tag";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { Settings } from "~/types.js";
 import type { ListTagsResponseItem } from "~/modules/FileManagerApiProvider/FileManagerApiContext/FileManagerApiContext.js";
