+++ packages/app-file-manager/src/modules/FileManagerRenderer/FileManagerView/formatFileItem.ts
@@ -1,4 +1,4 @@
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { FileManagerFileItem } from "@webiny/app-admin";
 
 /**
