+++ packages/app-file-manager/src/modules/FileManagerRenderer/FileManagerViewProvider/setSelection.ts
@@ -1,4 +1,4 @@
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import type { State } from "./state.js";
 
 interface GetSelectionParams {
