+++ packages/app-file-manager/src/modules/FileManagerRenderer/FileManagerViewProvider/useListFiles.ts
@@ -3,7 +3,7 @@ import { validateOrGetDefaultDbSort } from "@webiny/app-aco/sorting.js";
 import { useGetDescendantFolders } from "@webiny/app-aco";
 import type { ListMeta } from "@webiny/app-aco/types.js";
 import { useSecurity } from "@webiny/app-admin";
-import type { FileItem } from "@webiny/app-admin/types.js";
+import type { FileItem } from "~/types.js";
 import { useStateIfMounted } from "@webiny/app-admin";
 import type {
     Loading,
